pyproject.toml
[tool.poetry]
name = "langchain-course"
version = "0.1.0"
description = "A tutorial project about how to use the langchain framework to interact with the large language model."
authors = ["tongtie <tie0214@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_course"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.239"
openai = "^0.27.8"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
google-search-results = "^2.4.2"
duckduckgo-search = "^3.8.4"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "default"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
